modified the sequelize env files
