+++ .gitignore
@@ -3,3 +3,7 @@
 
 # environment variables
 .env
+
+# sequelize varaibles
+.config
+
